requirements.txt
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
schedule==1.2.2
pytz==2024.2
python-dotenv==1.0.1